Csv writer and simplification
